requirements.txt
pip==21.3.1
requests~=2.26.0
pytest~=6.2.5
djangorestframework~=3.12.4
Django~=3.2.9
celery~=5.2.1
djoser==2.1.0
psycopg2-binary==2.9.2
redis==4.0.2
flake8==4.0.1
isort==5.10.1
